+++ catalyst/callbacks/metrics/dice.py
@@ -19,19 +19,7 @@ if TYPE_CHECKING:
 
 
 class DiceCallback(BatchMetricCallback):
-    """Dice metric callback.
-
-    Args:
-        input_key: input key to use for iou calculation
-            specifies our ``y_true``
-        output_key: output key to use for iou calculation;
-            specifies our ``y_pred``
-        prefix: key to store in logs
-        eps: epsilon to avoid zero division
-        threshold: threshold for outputs binarization
-        activation: An torch.nn activation applied to the outputs.
-            Must be one of ``'none'``, ``'Sigmoid'``, ``'Softmax'``
-    """
+    """Dice metric callback."""
 
     def __init__(
         self,
@@ -53,6 +41,11 @@ class DiceCallback(BatchMetricCallback):
                 Must be one of ``'none'``, ``'Sigmoid'``, or ``'Softmax'``
             class_args: class names to display in the logs.
                 If None, defaults to indices for each class, starting from 0
+            **kwargs: key-value params to pass to the metric
+
+        .. note::
+            For `**kwargs` info, please follow
+            `catalyst.metrics.dice.dice` docs
         """
         metric_fn = wrap_metric_fn_with_activation(
             metric_fn=dice, activation=activation